Fix: Update offline messages by business hour
